create update_all_remotes.sh
This script automates cloning the latest origin of a bare repository to
a temprorary directory and then force pushing the latest changes into
that bare repository to immediately update the data available to users
via cgit. When run as a cronjob, this automates the process of making
sure all repos in cgit reflect the latest changes at the url set for
the git remote origin.

Signed-off-by: Joel Savitz <jsavitz@redhat.com>